Update 20220113 - CRUD Implementation
+ Added URL scheming on certain environments
* Finished Reservation CRUD
* Compressed statuses to be a single column
* Updated various data types for the database
* Updated calendar's start time and various toolkit button texts
* Updated header to properly position itself on different breakpoints
* Updated routing for various reservation actions

+++ database/migrations/2022_12_06_055129_create_menus_table.php
@@ -16,7 +16,7 @@ class CreateMenusTable extends Migration
 		Schema::create('menus', function (Blueprint $table) {
 			$table->id();
 			$table->string('name');
-			$table->double('price')->unsigned()->default(0);
+			$table->decimal('price', 12, 2)->unsigned()->default(0);
 			$table->time('duration')->default('01:00');
 			$table->softDeletes();
 			$table->timestamps();
